refactor users events controller and routes
