+++ db/seeds.rb
@@ -5,6 +5,9 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
+User.destroy_all
+Event.destroy_all
+Location.destroy_all
 
 User.create(name: 'Joe', email: 'joedoe@gmail.com')
 Event.create(starts_at: Time.now, ends_at: Time.now+5)
